align.col {soil.spec}	R Documentation

Combines two files and matches columns

Description

Combines two files with common data even when columns are not arranged in the same order.

Usage

align.col(x,y)

Arguments

x
First file to be used for combination with the second

y
The other file to be used to combine with the first

Details

Files comes with variables in different orders

Value

The original replicate spectra is returned together with averaged table. A graphical output of the averaged rows is provided and can be saved from the menu if necessary

x.y
combined file

Note

If used on different form of data other than IR, no spectrums are generated and a warning message is given which should be ignored. Also, ensure that column names have prefix; data tables obtained from read.spc should first be saved and then read back that way numeric column names will be prefixed automatically. User can control prefix appended to the wavenumbers by naming them explicitly.

Author(s)

Andrew Sila

Examples


## Not run: mir<-align.col(x,y)

[Package soil.spec version 0.2.0 Index]