average {soil.spec}	R Documentation

Averages data with replicate records

Description

Replicate data can be for any type of numeric values. An option to specify if data is spectra or non-spectra is required for this function to determine if to produce spectral signatures of averaged records or not incase of non-spectral data.

Usage


average(spec, type="")

Arguments

spec
Spec is a file with replicate data to be averaged, first column should include duplicated ids to identify rows to be averaged

type
Specify if file with replicate data has spectral data type or not

Details

Only numeric columns will be averaged, otherwise NA values will be returned

Value

The original replicate spectra is returned together with averaged table. A graphical output of the averaged rows is provided and can be saved from the menu if necessary

rep.spec
Replicated data before averaging

averaged
Data frame of the average

Note

1. If used on different form of data other than IR, no graphics for spectrums are generated

Note

2. Ensure that column names have prefix; data tables obtained from read.spc should first be saved and then read back that way numeric column names will be prefixed automatically. User can control prefix appended to the wavenumbers by naming them explicitly.

Author(s)

Andrew Sila

Examples


## Not run: spec<-read.csv(file.choose())
## Not run: average(spec,type="spectra")
## Not run: other<-read.csv(file.choose())
## Not run: average(other,type="")

[Package soil.spec version 0.2.0 Index]