explore.data {soil.spec}	R Documentation

Data summary

Description

Another way of getting data summary statitics and organized in a user friendly way- tabular form

Usage


explore.data(datafile)

Arguments

datafile

Details

This takes a table and returns the total number of complete cases (excluding missing values), mean,minimum, maximum and standard deviation values for numeric columns in a data table. The output is well formated for a quick and useful exploration view

Value

complete.cases
Number of complete observations

Average
Mean of each numeric column

Min
Minimum value of each numeric column

Max
Maximum value of each numeric column

Sd
Standard deviation value of each numeric column

Note

If all columns are non-numeric the complete cases will be returned only

Author(s)

Andrew Sila

Examples



## Not run: #Start by reading a file with data to be explored
	## Not run: soil<-read.csv(file.choose())
	## Not run: explore.data(soil)

[Package soil.spec version 0.2.0 Index]