spec.env {soil.spec}	R Documentation

Soil.spec package specific environmental variables

Description

Generates internal environment / package specific parameters and settings that can be later on passed to other functions.

Usage

spec.env(MIR = c(390, 7500), NIRS = c(3900, 12500), 
   NIRP = c(4000,10000), VISNIR1 = c(420, 960), 
   VISNIR2 = c(1020, 1770), 
   VISNIR3 = c(1830, 2480),
   icraf.htsxt = c(3578, 7497.964, 599.76), 
   icraf.alpha = c(2542, 3998.12872, 399.387991), 
   icraf.mpa = c(2307, 12493.2, 3598.69),  
   CO2.band = c(2350.8,2379.8), signif.digit = 5, 
   attributes = c("ORCCNS", "PHIHOX", "ALUM3S", 
   "ECAM3S", "EXKM3S", "EMGM3S", "ENAM3S", "EXB", 
   "NITCNS", "SNDLDF"), 
   mdnames = c("MID", "Instrument_name", "Instrument_URL", 
   "Laboratory_name", "Laboratory_contact", "Laboratory_URL",
   "Material_class", "Wavenumber_conversion", "Wavenlength_unit",
   "Location_error"), show.env = FALSE)

Arguments

MIR
"numeric", mid-infrared part of spectra

NIRS
"numeric", near infrared part of spectra

NIRP
"numeric", near infrared part of spectra

VISNIR1
"numeric", visible near infrared part of spectra

VISNIR2
"numeric", visible near infrared part of spectra

VISNIR3
"numeric", visible near infrared part of spectra

icraf.htsxt
"numeric", ICRAF MIR part of spectra

icraf.alpha
"numeric", ICRAF alpha-MIR part of spectra

icraf.mpa
"numeric", ICRAF VISNIR part of spectra

CO2.band
"numeric", wavenumbers for the CO2 band

signif.digit
"integer", default rounding system for absorbances

attributes
"character", standard target variables of interest: "ORCCNS" (Organic carbon content in soil estimated using the CNS elemental analyzer), "PHIHOX" (soil pH measured in water), "ALUM3S" (Aluminium concentration estimated using the Mehlich 3 solution), "ECAM3S" (exchangeable Calcium determined using using the Mehlich 3 solution), "EXKM3S" (exchangeable Potassium determined using using the Mehlich 3 solution), "EMGM3S" (exchangeable Magnesium determined using using the Mehlich 3 solution), "ENAM3S" (exchangeable Sodium determined using using the Mehlich 3 solution), "EXB" (sum of exchangeable bases), "NITCNS" (total nitrogen estimated using estimated using the CNS elemental analyzer), "SNDLDF" (weight percentage of particles 50-2 mm size determined using the laser diffraction method)

mdnames
"character", standard metadata names

show.env
"logical", specifies whether the output environment should be printed

Author(s)

Andrew Sila and Tomislav Hengl

Examples


# environmental variables:
spec.env()
get("mdnames", envir = spec.opts)
get("attributes", envir = spec.opts)

[Package soil.spec version 0.2.0 Index]